// shared/schema.d.ts
export declare const forgotPasswordSchema: z.ZodObject<{
    email: z.ZodString;
}, "strip", z.ZodTypeAny, {
    email: string;
}, {
    email: string;
}>;
export declare const resetPasswordSchema: z.ZodEffects<z.ZodObject<{
    token: z.ZodString;
    password: z.ZodString;
    confirmPassword: z.ZodString;
}, "strip", z.ZodTypeAny, {
    token: string;
    password: string;
    confirmPassword: string;
}, {
    token: string;
    password: string;
    confirmPassword: string;
}>, {
    token: string;
    password: string;
    confirmPassword: string;
}, {
    token: string;
    password: string;
    confirmPassword: string;
}>;